Revert "Merge v0.5.0 into master (#982)"
This reverts commit d42de35c678658b6b532221dd65ec8e9003a5433.

+++ metrics/cputime_nop.go
@@ -21,6 +21,6 @@ package metrics
 
 // getProcessCPUTime returns 0 on Windows as there is no system call to resolve
 // the actual process' CPU time.
-func getProcessCPUTime() float64 {
+func getProcessCPUTime() int64 {
 	return 0
 }
